Some bugfix for AWS

+++ tilecloud_chain/__init__.py
@@ -301,11 +301,12 @@ class TileGeneration:
                     (extent[2], extent[1]),
                 ))
 
-    def add_geom_filter(self):
+    def add_geom_filter(self, queue_store=None):
         if self.geom:
             self.ifilter(IntersectGeometryFilter(
                     grid=self.get_grid(),
-                    geom=self.geom))
+                    geom=self.geom,
+                    queue_store=queue_store))
 
     def add_metatile_splitter(self):
         store = MetaTileSplitterTileStore(
@@ -314,7 +315,13 @@ class TileGeneration:
             self.layer['meta_buffer'])
 
         if self.options.test > 0:
-            self.tilestream = store.get(self.tilestream)
+            def meta_get(tilestream):
+                for metatile in tilestream:
+                    substream = store.get((metatile,))
+                    for tile in substream:
+                        tile.metatile = metatile
+                        yield tile
+            self.tilestream = meta_get(self.tilestream)
         else:
             def safe_get(tilestream):
                 for metatile in tilestream:
@@ -514,7 +521,7 @@ class IntersectGeometryFilter(object):
 
         if not intersects and hasattr(tile, 'metatile'):
             tile.metatile.elapsed_togenerate -= 1
-            if tile.metatile.elapsed_togenerate == 0:
+            if tile.metatile.elapsed_togenerate == 0 and self.queue_store is not None:
                 self.queue_store.delete_one(tile.metatile)  # pragma: no cover
 
         return tile if intersects else None
